+++ src/service/file_service.rs
@@ -2,10 +2,13 @@ use regex::Regex;
 use std::fs::File;
 use std::path::Path;
 
+use crate::db;
+use crate::db::file_repository;
 use rocket::tokio::fs::create_dir;
+use rusqlite::Connection;
+use sha2::{Digest, Sha256};
 
-use crate::facade::file_facade::{delete_file_by_id, get_file_info_by_id, save_file_record};
-use crate::facade::{file_facade, folder_facade};
+use crate::facade::folder_facade;
 use crate::model::db::{FileRecord, Folder};
 use crate::model::request::file_requests::CreateFileRequest;
 use crate::model::response::file_responses::FileMetadataResponse;
@@ -103,7 +106,7 @@ async fn persist_save_file_to_folder(
         Ok(_) => {
             // since this is guaranteed to happen after the file is successfully saved, we can unwrap here
             let mut saved_file = File::open(&file_name).unwrap();
-            match file_facade::save_file_record(&file_name, &mut saved_file) {
+            match save_file_record(&file_name, &mut saved_file) {
                 Ok(id) => {
                     // file and folder are both in db, now link them
                     if let Err(_) = folder_facade::link_folder_to_file(id, folder.id.unwrap()) {
@@ -152,11 +155,24 @@ async fn persist_save_file(file_input: &mut CreateFileRequest<'_>) -> Result<u32
 }
 
 pub fn get_file(id: u32) -> Result<FileRecord, GetFileError> {
-    get_file_info_by_id(id)
+    let con = db::open_connection();
+    let result = match file_repository::get_by_id(id, &con) {
+        Ok(record) => Ok(record),
+        Err(error) if error == rusqlite::Error::QueryReturnedNoRows => Err(GetFileError::NotFound),
+        Err(error) => {
+            eprintln!(
+                "Failed to pull file info from database! Nested exception is: \n {:?}",
+                error
+            );
+            Err(GetFileError::DbFailure)
+        }
+    };
+    con.close().unwrap();
+    result
 }
 
 pub fn download_file(id: u32) -> Result<File, GetFileError> {
-    let res = file_facade::get_file_path(id);
+    let res = get_file_path(id);
     return if let Ok(path) = res {
         let path = format!("{}/{}", FILE_DIR, path);
         match File::open(path) {
@@ -169,12 +185,16 @@ pub fn download_file(id: u32) -> Result<File, GetFileError> {
 }
 
 pub fn delete_file(id: u32) -> Result<(), DeleteFileError> {
-    let file_path = match file_facade::get_file_path(id) {
+    let file_path = match get_file_path(id) {
         Ok(path) => format!("{}/{}", FILE_DIR, path),
         Err(e) if e == GetFileError::NotFound => return Err(DeleteFileError::NotFound),
         Err(_) => return Err(DeleteFileError::DbError),
     };
-    match delete_file_by_id(id) {
+    // now that we've determined the file exists, we can remove from the db
+    let con = db::open_connection();
+    let delete_result = delete_file_by_id_with_connection(id, &con);
+    con.close().unwrap();
+    match delete_result {
         Ok(_) => {
             match std::fs::remove_file(&file_path) {
                 Ok(()) => Ok(()),
@@ -186,4 +206,54 @@ pub fn delete_file(id: u32) -> Result<(), DeleteFileError> {
         }
         Err(e) => Err(e),
     }
+}
+
+// ==== private functions ==== \\
+fn save_file_record(name: &String, mut file: &mut File) -> Result<u32, String> {
+    let con = db::open_connection();
+    // hash the file TODO remove - we won't check uniqueness by file hash anymore
+    let mut hasher = Sha256::new();
+    std::io::copy(&mut file, &mut hasher).unwrap();
+    let hash = hasher.finalize();
+    // remove the './' from the file name
+    let begin_path_regex = Regex::new("\\.?(/.*/)+?").unwrap();
+    let mut formatted_name = begin_path_regex.replace(&name, "");
+    let hash = format!("{:x}", hash);
+    let file_record = FileRecord::from(formatted_name.to_mut().to_string(), hash);
+    let res = file_repository::save_file_record(&file_record, &con);
+    con.close().unwrap();
+    res
+}
+
+fn get_file_path(id: u32) -> Result<String, GetFileError> {
+    let con = db::open_connection();
+    let result = match file_repository::get_file_path(id, &con) {
+        Ok(path) => Ok(path),
+        Err(e) if e == rusqlite::Error::QueryReturnedNoRows => Err(GetFileError::NotFound),
+        Err(e) => {
+            eprintln!("Failed to get file path! Nested exception is: \n {:?}", e);
+            Err(GetFileError::DbFailure)
+        }
+    };
+    con.close().unwrap();
+    result
+}
+
+/// for use in folder_facade, so that way we don't create a new connection every time
+fn delete_file_by_id_with_connection(
+    id: u32,
+    con: &Connection,
+) -> Result<FileRecord, DeleteFileError> {
+    let result = match file_repository::delete_by_id(id, &con) {
+        Ok(record) => Ok(record),
+        Err(e) if e == rusqlite::Error::QueryReturnedNoRows => Err(DeleteFileError::NotFound),
+        Err(e) => {
+            eprintln!(
+                "Failed to delete file record from database! Nested exception is: \n {:?}",
+                e
+            );
+            Err(DeleteFileError::DbError)
+        }
+    };
+    return result;
 }
